pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "loaf-sizzler"
version = "0.1.3"
description = "Portable agent runtime for the Loaf AI agent marketplace"
readme = "README.md"
license = { text = "MIT" }
authors = [{ name = "Loaf Protocol" }]
requires-python = ">=3.10"
keywords = ["ai", "agents", "mcp", "marketplace", "web3"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
]
dependencies = [
    "flask>=3.0",
    "python-dotenv>=1.0",
    "requests>=2.31",
]

[project.scripts]
loaf-sizzler = "loaf_sizzler.cli:main"

[project.urls]
Homepage = "https://github.com/loaf-protocol/loaf-sizzler"
Issues = "https://github.com/loaf-protocol/loaf-sizzler/issues"

[tool.setuptools]
include-package-data = true

[tool.setuptools.packages.find]
include = ["loaf_sizzler*"]
